+++ inji-web-test/src/test/java/stepdefinitions/StepDef.java
@@ -4,17 +4,15 @@ package stepdefinitions;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-import org.apache.pdfbox.pdmodel.PDDocument;
-import org.apache.pdfbox.text.PDFTextStripper;
 import org.junit.Assert;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
-import pages.*;
+import pages.HelpPage;
+import pages.HomePage;
+import pages.SetNetwork;
 import utils.BaseTest;
 import utils.GlobelConstants;
 
-import java.io.*;
-import java.util.Base64;
 import java.util.Set;
 
 
@@ -25,15 +23,11 @@ public class StepDef {
     private GlobelConstants globelConstants;
     private HomePage homePage;
     private HelpPage helpPage;
-    private MosipCredentials mosipCredentials;
-    private SunbirdCredentials sunbirdCredentials;
     private SetNetwork setNetwork;
 
     public StepDef(BaseTest baseTest) {
         this.baseTest = baseTest;
         this.homePage = new HomePage(baseTest.getDriver());
-        this.sunbirdCredentials = new SunbirdCredentials(baseTest.getDriver());
-        this.mosipCredentials = new MosipCredentials(baseTest.getDriver());
         this.helpPage = new HelpPage(baseTest.getDriver());
         this.setNetwork = new SetNetwork();
     }
@@ -53,19 +47,16 @@ public class StepDef {
         Assert.assertEquals(pageTitle, pageTitle);
     }
 
-    //
     @Then("User verify that inji web logo is displayed")
     public void verifyInjiWebLogoIsDisplayed() throws InterruptedException {
         Assert.assertTrue(homePage.isLogoDisplayed());
     }
 
-    //
     @When("User clicks on the help button")
     public void clicksOnHelpButton() {
         homePage.ClickOnHelpForMobileBrowse();
         homePage.clickOnHelp();
     }
-
 
     @Given("Load application url {string}")
     public void load_application_url(String string) {
@@ -83,36 +74,6 @@ public class StepDef {
     @Then("User verify list of credential types displayed")
     public void user_verify_list_of_credential_types_displayed() {
         Assert.assertTrue(homePage.isListOfCredentialsTypeDisplayed());
-    }
-
-    @Then("User verify mock verifiable credential by e-signet displayed")
-    public void user_verify_mock_verifiable_credential_by_e_signet_displayed() {
-        Assert.assertTrue(mosipCredentials.isMockVerifiableCredentialDisplayed());
-    }
-
-    @When("User click on mock verifiable credential by e-signet button")
-    public void user_click_on_mock_verifiable_credential_by_e_signet_button() {
-        mosipCredentials.clickOnMockVerifiableCredential();
-    }
-
-    @When("User enter the  {string}")
-    public void user_enter_the(String string) {
-        mosipCredentials.enterVid(string);
-    }
-
-    @When("User click on getOtp button")
-    public void user_click_on_get_otp_button() {
-        mosipCredentials.clickOnGetOtpButton();
-    }
-
-    @When("User enter the otp {string}")
-    public void user_enter_the_otp(String otpString) {
-        try {
-            Thread.sleep(3000);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-        mosipCredentials.enterOtp(baseTest.getDriver(), otpString);
     }
 
     @When("User click on verify button")
@@ -135,58 +96,9 @@ public class StepDef {
         Assert.assertTrue(homePage.isMosipNationalIdDisplayed());
     }
 
-    @When("User click on mosip national id by e-signet button")
-    public void user_click_on_mosip_national_id_by_e_signet_button() {
-        mosipCredentials.clickOnMosipNationalId();
-    }
-
-    @Then("User click on sunbird cridentials button")
-    public void click_on_sunbird_cridentials_button() {
-        homePage.scrollDownByPage(baseTest.getDriver());
-        sunbirdCredentials.clickOnDownloadSunbird();
-    }
-
-    @Then("User verify sunbird cridentials button")
-    public void user_verify_sunbird_cridentials_button() {
-        Assert.assertTrue(sunbirdCredentials.isDownloadSunbirdCredentialsDisplayed());
-    }
-
-    @Then("User verify sunbird rc insurance verifiable credential displayed")
-    public void user_verify_sunbird_rc_insurance_verifiable_credential_displayed() {
-        Assert.assertTrue(sunbirdCredentials.isSunbirdInsuranceDisplayed());
-
-    }
-
-    @Then("User click on sunbird rc insurance verifiable credential button")
-    public void user_click_on_sunbird_rc_insurance_verifiable_credential_button() {
-        sunbirdCredentials.clickOnSunbirdInsurance();
-    }
-
     @Then("User click on {string} button")
     public void user_click_on_button(String string) {
 
-    }
-
-    @Then("User enter the policy number {string}")
-    public void user_enter_the_policy_number(String string) throws InterruptedException {
-        Thread.sleep(3000);
-        sunbirdCredentials.enterPolicyNumer(string);
-
-    }
-
-    @Then("User enter the full name  {string}")
-    public void user_enter_the_full_name(String string) {
-        sunbirdCredentials.enterFullName(string);
-    }
-
-    @Then("User enter the date of birth {string}")
-    public void user_enter_the_date_of_birth(String string) {
-        sunbirdCredentials.selectDateOfBirth();
-    }
-
-    @Then("User click on login button")
-    public void user_click_on_login_button() {
-        sunbirdCredentials.clickOnLogin();
     }
 
     @Then("User search the issuers with {string}")
@@ -197,16 +109,6 @@ public class StepDef {
             throw new RuntimeException(e);
         }
         homePage.enterIssuersInSearchBox(string);
-    }
-
-    @Then("User verify life Insurance displayed")
-    public void user_verify_life_insurance_displayed() {
-        Assert.assertTrue(sunbirdCredentials.isLifeInceranceDisplayed());
-    }
-
-    @Then("User click on life Insurance button")
-    public void user_click_on_life_insurance_button() {
-        sunbirdCredentials.clickOnLifeInsurance();
     }
 
     @Then("User verify go home button")
@@ -225,16 +127,6 @@ public class StepDef {
 //		Assert.assertTrue(homePage.isBackButtonDisplayed());
     }
 
-    @When("User verify login page lables")
-    public void user_verify_login_page_lables() {
-        Assert.assertTrue(mosipCredentials.isLoginPageLableDisplayed());
-    }
-
-    @When("User verify vid input box header")
-    public void user_verify_vid_input_box_header() {
-        Assert.assertTrue(mosipCredentials.isVidInputBoxHeaderDisplayed());
-    }
-
     @Then("User verify that langauge button is displayed")
     public void verify_that_langauge_button_is_displayed() {
         Assert.assertTrue(homePage.isLanguageDisplayed());
@@ -248,68 +140,6 @@ public class StepDef {
     @Then("User Verify the no issuer found message")
     public void user_verify_the_no_issuer_found_message() {
         Assert.assertTrue(homePage.isNoIssuerFoundMessageDisplayed());
-    }
-
-	@Then("User verify pdf is downloaded")
-	public String user_verify_pdf_is_downloaded() throws IOException {
-        baseTest.getJse().executeScript("browserstack_executor: {\"action\": \"fileExists\", \"arguments\": {\"fileName\": \"" + baseTest.PdfNameForMosip + "\"}}");
-        baseTest.getJse().executeScript("browserstack_executor: {\"action\": \"getFileProperties\", \"arguments\": {\"fileName\": \"" + baseTest.PdfNameForMosip + "\"}}");
-
-        String base64EncodedFile = (String) baseTest.getJse().executeScript("browserstack_executor: {\"action\": \"getFileContent\", \"arguments\": {\"fileName\": \"" + baseTest.PdfNameForMosip + "\"}}");
-        byte[] data = Base64.getDecoder().decode(base64EncodedFile);
-        OutputStream stream = new FileOutputStream(baseTest.PdfNameForMosip);
-        stream.write(data);
-
-        System.out.println(stream);
-        stream.close();
-
-        File pdfFile = new File(System.getProperty("user.dir") + "/" + baseTest.PdfNameForMosip);
-        PDDocument document = PDDocument.load(pdfFile);
-
-        PDFTextStripper stripper = new PDFTextStripper();
-        String text = stripper.getText(document);
-        return text;
-    }
-
-
-    @Then("User verify pdf is downloaded for Insurance")
-    public String user_verify_pdf_is_downloaded_for_insurance() throws IOException {
-        baseTest.getJse().executeScript("browserstack_executor: {\"action\": \"fileExists\", \"arguments\": {\"fileName\": \"" + baseTest.PdfNameForInsurance + "\"}}");
-        baseTest.getJse().executeScript("browserstack_executor: {\"action\": \"getFileProperties\", \"arguments\": {\"fileName\": \"" + baseTest.PdfNameForInsurance + "\"}}");
-
-        String base64EncodedFile = (String) baseTest.getJse().executeScript("browserstack_executor: {\"action\": \"getFileContent\", \"arguments\": {\"fileName\": \"" + baseTest.PdfNameForInsurance + "\"}}");
-        byte[] data = Base64.getDecoder().decode(base64EncodedFile);
-        OutputStream stream = new FileOutputStream(baseTest.PdfNameForInsurance);
-        stream.write(data);
-
-        System.out.println(stream);
-        stream.close();
-
-        File pdfFile = new File(System.getProperty("user.dir") + "/" + baseTest.PdfNameForInsurance);
-        PDDocument document = PDDocument.load(pdfFile);
-
-        PDFTextStripper stripper = new PDFTextStripper();
-        String text = stripper.getText(document);
-        return text;
-    }
-    @Then("User verify policy number input box header")
-    public void user_verify_policy_number_input_box_header() {
-        Assert.assertTrue(sunbirdCredentials.isEnterPolicyNumberHeaderDisplayed());
-    }
-
-    @Then("User verify full name input box header")
-    public void user_verify_full_name_input_box_header() {
-        Assert.assertTrue(sunbirdCredentials.isEnterFullNameHeaderDisplayed());
-    }
-
-    @Then("User verify date of birth input box header")
-    public void user_verify_date_of_birth_input_box_header() {
-        Assert.assertTrue(sunbirdCredentials.isEnterDOBHeaderDisplayed());
-    }
-
-    @Then("User verify authentication failed message")
-    public void user_verify_authentication_failed_message() {
-        Assert.assertTrue(sunbirdCredentials.isAuthenticationFailedDisplayed());
     }
 
     @Then("User click on arabic langauge")
@@ -451,16 +281,6 @@ public class StepDef {
         baseTest.getDriver().navigate().back();
     }
 
-    @Then("User verify Vehicle Insurance displayed")
-    public void user_verify_vehicle_insurance_displayed() {
-        Assert.assertTrue(sunbirdCredentials.isVehicleInsuranceDisplayed());
-    }
-
-    @Then("User click on Vehicle Insurance button")
-    public void user_click_on_vehicle_insurance_button() {
-        sunbirdCredentials.clickOnVehicleInsurance();
-    }
-
     @Then("User open new tab")
     public void user_open_new_tab() {
         ((JavascriptExecutor) baseTest.getDriver()).executeScript("window.open('" + baseTest.url + "')");
@@ -491,11 +311,6 @@ public class StepDef {
 
     @Then("user verify the page after Refresh")
     public void user_verify_the_page_after_refresh() {
-    }
-
-    @Then("User verify downloading in progress text")
-    public void user_VerifyDownloadingInProgressDisplaed() {
-        Assert.assertTrue(mosipCredentials.isDownloadingDescriptionTextDisplayed());
     }
 
     @When("User verify the FAQ header and its description")
